Add support for towncrier
Instead of updating the news.rst file manually each month
it would be far easier to add a news fragment for each feature
or change and let towncrier generate the news.rst file for you.

To do so, pyproject.toml needed to include some information about
the project so that towncrier can be configured properly. A fragments
directory was also included to be populated with news fragements.